feat: convert all monetary values from decimal to long (cents)

+++ PagueVeloz.TransactionEngine/src/PagueVeloz.Application/Commands/Accounts/OpenAccountCommand.cs
@@ -3,4 +3,4 @@ using PagueVeloz.Application.DTOs.Accounts.Responses;
 
 namespace PagueVeloz.Application.Commands.Accounts;
 
-public record OpenAccountCommand(Guid CustomerId, decimal CreditLimit = 0m) : IRequest<AccountResponse>;
+public record OpenAccountCommand(Guid CustomerId, long CreditLimit = 0) : IRequest<AccountResponse>;
